# apps/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CustomerOrder" : {
        "properties" : {
          "brand" : {
            "type" : "string"
          },
          "confirmedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "depot" : {
            "type" : "string"
          },
          "district" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "isoWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "isoYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "orderDate" : {
            "format" : "date",
            "type" : "string"
          },
          "outletId" : {
            "type" : "string"
          },
          "placedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "placedBy" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "planningDate" : {
            "description" : "Planning run the order belongs to; later than orderDate after a published deferral",
            "format" : "date",
            "type" : "string"
          },
          "ref" : {
            "type" : "string"
          },
          "sourceDaysSinceServed" : {
            "description" : "Days since last served, supplied with imported scenario data",
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "sourceDeferredYesterday" : {
            "description" : "Deferred-yesterday flag supplied with imported scenario data",
            "type" : [ "boolean", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "tempRequirement" : {
            "type" : "string"
          },
          "units" : {
            "format" : "int32",
            "type" : "integer"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          },
          "volumeM3" : {
            "type" : "number"
          },
          "weightKg" : {
            "type" : "number"
          }
        },
        "required" : [ "outletId", "ref" ],
        "type" : "object"
      },
      "CutoffInfo" : {
        "properties" : {
          "cutoffLocalTime" : {
            "type" : "string"
          },
          "nextCutoffAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "nextDeliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "open" : {
            "type" : "boolean"
          },
          "secondsRemaining" : {
            "format" : "int64",
            "type" : "integer"
          },
          "serverNow" : {
            "format" : "date-time",
            "type" : "string"
          },
          "timeZone" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OrderEstimate" : {
        "properties" : {
          "allowed" : {
            "description" : "Whether this outlet may place orders of this temperature",
            "type" : "boolean"
          },
          "basisOrders" : {
            "description" : "How many past orders the estimate is based on",
            "format" : "int64",
            "type" : "integer"
          },
          "chilledAllowed" : {
            "description" : "Whether this outlet may place chilled orders at all",
            "type" : "boolean"
          },
          "deliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "tempRequirement" : {
            "type" : "string"
          },
          "units" : {
            "format" : "int32",
            "type" : "integer"
          },
          "volumeM3" : {
            "type" : [ "number", "null" ]
          },
          "weightKg" : {
            "description" : "Null when there are no past orders to estimate from",
            "type" : [ "number", "null" ]
          },
          "windowClose" : {
            "type" : "string"
          },
          "windowOpen" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "OrderPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/CustomerOrder"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PlaceOrderRequest" : {
        "properties" : {
          "expectedDeliveryDate" : {
            "format" : "date",
            "type" : "string"
          },
          "tempRequirement" : {
            "minLength" : 1,
            "type" : "string"
          },
          "units" : {
            "format" : "int32",
            "maximum" : 100000,
            "minimum" : 1,
            "type" : "integer"
          },
          "volumeM3" : {
            "maximum" : 9999999.999,
            "minimum" : 0.001,
            "type" : "number"
          },
          "weightKg" : {
            "maximum" : 9.999999999E7,
            "minimum" : 0.01,
            "type" : "number"
          }
        },
        "required" : [ "tempRequirement", "units", "volumeM3", "weightKg" ],
        "type" : "object"
      },
...
      "ReceiptDeliveryRow" : {
        "properties" : {
          "deliveredAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "deliveredUnits" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "driverName" : {
            "type" : [ "string", "null" ]
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderRef" : {
            "type" : "string"
          },
          "outcome" : {
            "description" : "DELIVERED, PARTIAL or FAILED, once recorded",
            "type" : [ "string", "null" ]
          },
          "phase" : {
            "description" : "PENDING (planned or loading), IN_DELIVERY or DELIVERED (the driver recorded an outcome)",
            "type" : "string"
          },
          "planDate" : {
            "format" : "date",
            "type" : "string"
          },
          "plannedArrival" : {
            "type" : [ "string", "null" ]
          },
          "receipt" : {
            "description" : "NONE, CONFIRMED, DISPUTED (open) or RESOLVED",
            "type" : "string"
          },
          "tempRequirement" : {
            "type" : "string"
          },
          "tripIndex" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "units" : {
            "format" : "int32",
            "type" : "integer"
          },
          "vehicleId" : {
            "type" : [ "string", "null" ]
          },
          "windowClose" : {
            "type" : [ "string", "null" ]
          },
          "windowOpen" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "ReceiptDiscrepancy" : {
        "properties" : {
          "affectedUnits" : {
            "format" : "int32",
            "type" : "integer"
          },
          "decision" : {
            "description" : "CREDIT, REPLACEMENT or NO_ACTION",
            "type" : [ "string", "null" ]
          },
          "decisionNote" : {
            "type" : [ "string", "null" ]
          },
          "deliveredAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "deliveredUnits" : {
            "format" : "int32",
            "type" : "integer"
          },
          "depot" : {
            "type" : "string"
          },
          "driverName" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          },
          "note" : {
            "type" : [ "string", "null" ]
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderRef" : {
            "type" : "string"
          },
          "outletId" : {
            "type" : "string"
          },
          "receiptId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reportedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reportedByName" : {
            "type" : "string"
          },
          "resolvedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "resolvedByName" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "description" : "OPEN or RESOLVED",
            "type" : "string"
          },
          "vehicleId" : {
            "type" : [ "string", "null" ]
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "StoreHome" : {
        "description" : "The store manager's home summary. Every count is computed on the server for the signed-in outlet.",
        "properties" : {
          "brand" : {
            "type" : "string"
          },
          "completedOrders" : {
            "description" : "Orders delivered in full or part, or already received",
            "format" : "int64",
            "type" : "integer"
          },
          "depot" : {
            "type" : "string"
          },
          "district" : {
            "type" : "string"
          },
          "latestOpenIssue" : {
            "$ref" : "#/components/schemas/ReceiptDiscrepancy",
            "description" : "The most recently reported open issue",
            "type" : "null"
          },
          "nextDelivery" : {
            "$ref" : "#/components/schemas/ReceiptDeliveryRow",
            "description" : "The earliest delivery still to arrive",
            "type" : "null"
          },
          "openIssues" : {
            "description" : "Reported issues the dispatcher has not resolved",
            "format" : "int32",
            "type" : "integer"
          },
          "openOrders" : {
            "description" : "Orders confirmed, planned, loaded, in transit or deferred",
            "format" : "int64",
            "type" : "integer"
          },
          "outletId" : {
            "type" : "string"
          },
          "pendingDeliveries" : {
            "description" : "Deliveries still planned, loading or on the road",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "StoreNotification" : {
        "description" : "One thing that happened to the outlet's orders, newest first. Built from stored events; there is no read state.",
        "properties" : {
          "at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "date" : {
            "description" : "The delivery or planning day the event is about, when there is one",
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "kind" : {
            "description" : "ORDER_SUBMITTED, ORDER_DEFERRED, ORDER_DISPATCHED, ORDER_DELIVERED, RECEIPT_CONFIRMED, ISSUE_REPORTED or ISSUE_RESOLVED",
            "type" : "string"
          },
          "orderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderRef" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "StoreOrderBoard" : {
        "properties" : {
          "counts" : {
            "$ref" : "#/components/schemas/StoreOrderCounts"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/StoreOrderRow"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "StoreOrderCounts" : {
        "description" : "How many of the outlet's orders sit in each chip. Independent of the search and the selected chip.",
        "properties" : {
          "all" : {
            "format" : "int32",
            "type" : "integer"
          },
          "deferred" : {
            "format" : "int32",
            "type" : "integer"
          },
          "delivered" : {
            "format" : "int32",
            "type" : "integer"
          },
          "inDelivery" : {
            "format" : "int32",
            "type" : "integer"
          },
          "issue" : {
            "format" : "int32",
            "type" : "integer"
          },
          "planned" : {
            "format" : "int32",
            "type" : "integer"
          },
          "submitted" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "StoreOrderRow" : {
        "description" : "One order on the store manager's Orders screen, with the group its chip belongs to.",
        "properties" : {
          "deliveryDate" : {
            "description" : "The planned delivery day once known, else the planning run the order sits in",
            "format" : "date",
            "type" : "string"
          },
          "group" : {
            "description" : "SUBMITTED, PLANNED, IN_DELIVERY, DELIVERED, ISSUE, DEFERRED or CANCELLED",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "orderDate" : {
            "format" : "date",
            "type" : "string"
          },
          "placedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "plannedArrival" : {
            "type" : [ "string", "null" ]
          },
          "planningDate" : {
            "format" : "date",
            "type" : "string"
          },
          "ref" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "tempRequirement" : {
            "type" : "string"
          },
          "units" : {
            "format" : "int32",
            "type" : "integer"
          },
          "volumeM3" : {
            "type" : "number"
          },
          "weightKg" : {
            "type" : "number"
          },
          "windowClose" : {
            "type" : [ "string", "null" ]
          },
          "windowOpen" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "session" : {
        "in" : "cookie",
        "name" : "WP_SESSION",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "title" : "TechTrithalon — Waypoint Operations API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/store/cutoff" : {
      "get" : {
        "operationId" : "cutoff",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CutoffInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-order-controller" ]
      }
    },
...
    "/api/v1/store/home" : {
      "get" : {
        "operationId" : "home",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreHome"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-receipt-controller" ]
      }
    },
    "/api/v1/store/issues" : {
      "get" : {
        "operationId" : "issues",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ReceiptDiscrepancy"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-receipt-controller" ]
      }
    },
    "/api/v1/store/notifications" : {
      "get" : {
        "operationId" : "notifications",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 30,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/StoreNotification"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-receipt-controller" ]
      }
    },
    "/api/v1/store/order-board" : {
      "get" : {
        "operationId" : "orderBoard",
        "parameters" : [ {
          "in" : "query",
          "name" : "group",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreOrderBoard"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-receipt-controller" ]
      }
    },
    "/api/v1/store/order-estimate" : {
      "get" : {
        "operationId" : "estimate",
        "parameters" : [ {
          "in" : "query",
          "name" : "temp",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "units",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderEstimate"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-order-controller" ]
      }
    },
    "/api/v1/store/orders" : {
      "get" : {
        "operationId" : "orders",
        "parameters" : [ {
          "in" : "query",
          "name" : "date",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "orderDate",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "asc",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 50,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderPage"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-order-controller" ]
      },
      "post" : {
        "operationId" : "place",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlaceOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerOrder"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-order-controller" ]
      }
    },
    "/api/v1/store/orders/{id}" : {
      "get" : {
        "operationId" : "order",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerOrder"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "session" : [ ]
        } ],
        "tags" : [ "store-order-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}